requirements.txt
deep_sort_realtime
matplotlib
numpy
opencv_python
pandas
torch
ultralytics
python-dotenv